+++ packages/node/src/configure/NodeConfig.ts
@@ -102,7 +102,6 @@ export class NodeConfig implements IConfig {
   get debug(): boolean {
     return this._config.debug;
   }
-
   get preferRange(): boolean {
     return this._config.preferRange;
   }