+++ packages/node/src/indexer/dictionary.service.ts
@@ -12,6 +12,8 @@ import { SubqlEventFilter, SubqlCallFilter, MetaData } from '@subql/common';
 import fetch from 'node-fetch';
 import { SubqueryProject } from '../configure/project.model';
 import { getLogger } from '../utils/logger';
+import { profiler } from '../utils/profiler';
+import { getYargsOption } from '../yargs';
 import { ProjectIndexFilters } from './types';
 
 export type Dictionary = {
@@ -21,6 +23,7 @@ export type Dictionary = {
   // specVersions: number[];
 };
 const logger = getLogger('dictionary');
+const { argv } = getYargsOption();
 
 @Injectable()
 export class DictionaryService implements OnApplicationShutdown {
@@ -39,6 +42,8 @@ export class DictionaryService implements OnApplicationShutdown {
    * @param batchSize
    * @param indexFilters
    */
+
+  @profiler(argv.profiler)
   async getDictionary(
     startBlock: number,
     queryEndBlock: number,