+++ packages/node/src/indexer/events.ts
@@ -10,6 +10,8 @@ export enum IndexerEvent {
   BlockQueueSize = 'block_queue_size',
   BlocknumberQueueSize = 'blocknumber_queue_size',
   NetworkMetadata = 'network_metadata',
+  UsingDictionary = 'using_dictionary',
+  SkipDictionary = 'skip_dictionary',
 }
 
 export interface ProcessBlockPayload {