+++ packages/node/src/yargs.ts
@@ -47,6 +47,12 @@ export function getYargsOption() {
       type: 'boolean',
       default: false,
     },
+    profiler: {
+      demandOption: false,
+      describe: 'Show profiler information to console output',
+      type: 'boolean',
+      default: false,
+    },
     'network-endpoint': {
       demandOption: false,
       type: 'string',