+++ packages/node/CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Added
+- add profiler to monitoring indexer performance
+- add metrics to listen status of using dictionary and number of times it been skipped.
 
 ## [0.17.2] - 2021-07-01
 Upgrade priority: High.